build fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@55052 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/testcase/tstHelp.h
@@ -37,7 +37,7 @@ RT_C_DECLS_END
         if (off != RT_OFFSETOF(type, m)) \
         { \
             printf("error! %#010x %s  Off by %d!! (expected off=%#x)\n", \
-                   RT_OFFSETOF(type, m), #type "." #m, off - RT_OFFSETOF(type, m), off); \
+                   RT_OFFSETOF(type, m), #type "." #m, off - RT_OFFSETOF(type, m), (int)off); \
             rc++; \
         } \
         /*else */ \
